Removed call to base.Update() in Bullet.Update()
It appears Glib.Xna.SpriteLib.Sprite.Update() is calling
Mouse.GetState() for every sprite. This is very expensive; I will open a
ticket on Glib project for this issue. Meanwhile, these calls amount to
over 70% of overall update time for Bullet class, and cause lag. This is
a temporary fix, until the bug in Glib is resolved.