Both ships are now drawn on both screens
- Correct ship type and tier are drawn on both host and remote gamer
screens
- Position is not maintained, even on init
- Both ships are still "Ally" player type on remote screen - if changed
manually in debugger, then it's possible to kill opposing player. Need
to set proper ship type - not what comes over, necessarily, as all
might've been instantiated as "Ally"; or fix the instantiation.
- We need to overhaul the networking code, starting with marking up
which functions are called when, and whether or not they are called by
hosting or joining system.